Move sub-configs to garget.rs
